Create RouterWrapper to update user state
