+++ src/main.tsx
@@ -2,6 +2,7 @@ import { History } from 'history';
 import * as React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { AppProvider } from "./context/context";
+import RouterWrapper from "./RouterWrapper";
 import Routes from './routes';
 
 interface AllProps {
@@ -13,7 +14,9 @@ const Main = (props: AllProps) => {
   return (
     <AppProvider>
       <BrowserRouter>
-        <Routes />
+        <RouterWrapper>
+          <Routes />
+        </RouterWrapper>
       </BrowserRouter>
     </AppProvider>
   );
